refactor : User.java
- User 클래스를 추가하여 기능 분리.
