+++ src/main/java/vendingmachine/Application.java
@@ -38,43 +38,11 @@ public class Application {
         return purchaseProductObject;
     }
 
-    public static void purchaseProduct(ArrayList<Product> products) throws IllegalArgumentException {
-        System.out.println("구매할 상품명을 입력해 주세요.");
-        String purchaseProductName = Console.readLine();
-        Product purchaseProductObject = findPurchaseProduct(products, purchaseProductName);
-        productNotFound(purchaseProductObject);
-        productAmountNotEnough(purchaseProductObject);
-        userMoneyNotEnough(purchaseProductObject);
-        purchaseSuccess(purchaseProductObject);
-    }
-
-    public static void purchaseSuccess(Product product){
-        product.sold();
-        USER_MONEY -= product.getPrice();
-    }
 
 
 
 
-    public static void insertMoneyByUser(){
-        String userMoney = "";
-        while(userMoney.isEmpty()){
-            try{
-                userMoney = userMoneyValidation();
-                USER_MONEY = Integer.parseInt(userMoney);
-            }catch (IllegalArgumentException e){
-                System.out.println("[ERROR] 금액은 숫자여야 합니다.");
-            }
-        }
-    }
 
-    public static String userMoneyValidation(){
-        System.out.println("투입 금액을 입력해 주세요.");
-        String userMoney = Console.readLine();
-        String regex = "^[0-9]*$";
-        if (!Pattern.matches(regex, userMoney)) throw new IllegalArgumentException();
-        return userMoney;
-    }
 
 
     public static ArrayList<Product> createProduct(ArrayList<String> products){
